Refactor views and filenames
