+++ views/user-profile/layout.pug
@@ -1,0 +1,26 @@
+doctype html
+html(lang="en")
+  head
+    title= title
+    link(rel="stylesheet", href="/css/styles.css")
+  body
+    header.navbar
+      a(href="/")
+        h1.title Blog with Express
+      nav
+        ul
+          li
+            a.navbar-button(href="/create") + Create Post
+          li
+            a.navbar-button(href="/stats") Post Stats
+      .navbar-account
+        .navbar-account-button
+          a.login-button(href="/login") Login
+          div /
+          a.register-button(href="/register") Register
+        a(href="/profile")
+          img(class="profile-icon", src=profileIcon, alt='Profile Picture')
+    main
+      h1= title
+      block content
+
